requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==26.1.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
daff==1.4.2
dbt-adapters==1.9.0
dbt-common==1.12.0
dbt-core==1.8.0
dbt-duckdb==1.8.0
dbt-extractor==0.6.0
dbt-semantic-interfaces==0.5.1
deepdiff==7.0.1
dlt==1.24.0
duckdb==1.5.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
giturlparse==0.14.0
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-secret-manager==2.27.0
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.62.3
humanize==4.15.0
idna==3.11
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
jsonpath-ng==1.7.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
Logbook==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.20
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
ordered-set==4.1.0
orjson==3.11.8
packaging==26.0
parsedatetime==2.6
pathspec==0.12.1
pathvalidate==3.3.1
pendulum==3.2.0
pluggy==1.6.0
ply==3.11
proto-plus==1.27.2
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requirements-parser==0.13.0
rich==14.3.3
rich-argparse==1.7.2
rpds-py==0.30.0
semver==3.0.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
sqlglot==30.3.0
sqlparse==0.5.5
tenacity==9.1.4
text-unidecode==1.3
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
zipp==3.23.0
